fix - proxy refactor user_api_key_auth
